requirements.txt
awscrt==0.20.0
awsiotsdk==1.20.0
python-dotenv==1.0.0
